requirements.txt
langchain
langchain-community
langchain-openai
chromadb
pymongo
pypdf
tiktoken
python-dotenv
